execution/commitment, db/state, node: remove ConcurrentHexPatriciaTrie PoC
The concurrent commitment trie was a proof-of-concept, superseded by the
parallel (ParallelPatriciaHashed) and streaming (StreamingCommitter)
commitment paths now on main. Remove the PoC trie, its
--experimental.concurrent-commitment flag and config wiring, and its
concurrent-exclusive tests.

- drop ConcurrentPatriciaHashed (+ ParallelHashSort/CanDoConcurrentNext),
  VariantConcurrentHexPatricia, statecfg.ExperimentalConcurrentCommitment and
  the --experimental.concurrent-commitment flag end to end
- relocate the shared mountTo primitive into parallel_mount.go and the shared
  nibble/address test helpers into nibble_addr_test.go
- backtester paraTrie now selects VariantParallelHexPatricia
- update the parallel-patricia-hashed design doc

Sequential / parallel / streaming commitment paths are unchanged;
SetConcurrentCommitment / IsConcurrentCommitment are kept (used by ModeParallel).
