+++ db/state/execctx/domain_shared.go
@@ -141,13 +141,11 @@ func PickTrieVariant() commitment.TrieVariant {
 	switch {
 	// Selecting more than one experimental-commitment flag is a misconfiguration;
 	// they are alternative paths. Streaming overlaps folding with execution, so it
-	// wins over parallel, which in turn parallelizes deeper than concurrent.
+	// wins over parallel.
 	case statecfg.ExperimentalStreamingCommitment:
 		return commitment.VariantStreamingHexPatricia
 	case statecfg.ExperimentalParallelCommitment:
 		return commitment.VariantParallelHexPatricia
-	case statecfg.ExperimentalConcurrentCommitment:
-		return commitment.VariantConcurrentHexPatricia
 	}
 	return commitment.VariantHexPatriciaTrie
 }
