+++ db/state/statecfg/state_schema.go
@@ -187,17 +187,14 @@ func (s *SchemaGen) GetBlockIdxFilesCfg(name string) BlockIdxFilesCfg {
 	return v
 }
 
-var ExperimentalConcurrentCommitment = false // set true to use concurrent commitment by default
-
 // ExperimentalParallelCommitment toggles the ParallelPatriciaHashed trie path
 // (commitment.ModeParallel + VariantParallelHexPatricia). Default false.
-// Takes precedence over ExperimentalConcurrentCommitment when both are set.
 var ExperimentalParallelCommitment = false
 
 // ExperimentalStreamingCommitment toggles the StreamingCommitter trie path
 // (commitment.ModeParallel + VariantStreamingHexPatricia), which overlaps
 // commitment folding with execution. Default false. Takes precedence over
-// ExperimentalParallelCommitment and ExperimentalConcurrentCommitment.
+// ExperimentalParallelCommitment.
 var ExperimentalStreamingCommitment = false
 
 var Schema = SchemaGen{
