+++ node/cli/default_flags.go
@@ -261,7 +261,6 @@ var DefaultFlags = []cli.Flag{
 
 	&utils.GDBMeFlag,
 
-	&utils.ExperimentalConcurrentCommitmentFlag,
 	&utils.ExperimentalParallelCommitmentFlag,
 	&utils.ExperimentalStreamingCommitmentFlag,
