+++ node/ethconfig/config.go
@@ -299,13 +299,12 @@ type Sync struct {
 	LoopBlockLimit             uint
 	ParallelStateFlushing      bool
 
-	ChaosMonkey                      bool
-	AlwaysGenerateChangesets         bool
-	MaxReorgDepth                    uint64
-	KeepExecutionProofs              bool
-	ExperimentalConcurrentCommitment bool
-	ExperimentalParallelCommitment   bool
-	ExperimentalStreamingCommitment  bool
-	PersistReceiptsCacheV2           bool
-	SnapshotDownloadToBlock          uint64 // exclusive [0,toBlock)
+	ChaosMonkey                     bool
+	AlwaysGenerateChangesets        bool
+	MaxReorgDepth                   uint64
+	KeepExecutionProofs             bool
+	ExperimentalParallelCommitment  bool
+	ExperimentalStreamingCommitment bool
+	PersistReceiptsCacheV2          bool
+	SnapshotDownloadToBlock         uint64 // exclusive [0,toBlock)
 }
